+++ decomp/ST.exe/functions/00594B90/decomp.c
@@ -29,11 +29,10 @@ FSGSTy::OutChSlProc(FSGSTy *this,AnonShape_GLOBAL_0080759C_9638EF10 *param_1,und
     errorCode = Library::MSVCRT::__setjmp3(local_50.jumpBuffer,0);
     puVar2 = local_8;
     if (errorCode == 0) {
-      Library::DKW::DDX::FUN_006c68f0
-                ((int)param_1,param_4,param_5,param_6,param_7,(int)(local_8 + 1));
+      Library::DKW::DDX::FUN_006c68f0(param_1,param_4,param_5,param_6,param_7,(int)(local_8 + 1));
       FUN_006c6850((int)param_1,param_4,param_5,param_6,param_7,*puVar2);
       Library::DKW::DDX::FUN_006c6660
-                ((int)param_1,param_4 + -0x15 + param_6,param_5 + 0x26,7,param_7 - 0x62,3,0);
+                (param_1,param_4 + -0x15 + param_6,param_5 + 0x26,7,param_7 - 0x62,3,0);
       Library::DKW::DDX::FUN_006c6470
                 ((int)param_1,param_4 + -0x15 + param_6,param_5 + 0x26,(undefined4 *)0x7,
                  param_7 + -0x62,3,*puVar2);
